Move soruces to Params, improve docstrings
