+++ src/params/sources/Sources.ts
@@ -1,0 +1,32 @@
+import TextSource from "./TextSource";
+import ImageSource from "./ImageSource";
+
+/**
+ * @memberOf Params
+ * @namespace Sources
+ */
+
+
+/**
+ * @memberOf Params.Sources
+ * @param {string} publicID
+ * @description Returns an instance of an ImageSource
+ * @returns {Params.Sources.ImageSource}
+ */
+function image(publicID: string): ImageSource {
+  return new ImageSource(publicID);
+}
+
+
+/**
+ * @memberOf Params.Sources
+ * @param someText {string}
+ * @description Returns an instance of a TextSource
+ * @returns {Params.Sources.TextSource}
+ */
+function text(someText: string): TextSource {
+  return new TextSource(someText);
+}
+
+export default {ImageSource, TextSource, image, text};
+export {ImageSource, TextSource, image, text};
